Runs correlation analysis pipeline

Runs correlation analysis pipeline

correlation_analysis(
  dataset_name,
  DEouts,
  celltype_correspondence,
  pvals = c(0.05, 0.025, 0.01, 0.001, 1e-04),
  data_names = "placeholder",
  alphaval = 0.25,
  N_randperms = 5,
  N_subsets = 5,
  sex_DEGs = FALSE,
  fontsize_yaxislabels = 12,
  fontsize_yaxisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  fontsize_facet_labels = 9,
  output_path = getwd()
)

Arguments

dataset_name

name of the dataset used to select significant DEGs from (specified as a string, name as in DEouts)

DEouts

a list containing outputs of DGE analysis (as returned/optionally saved by DGE_analysis) for datasets to be used in the correlation analysis

celltype_correspondence

list of different names specifying each cell type

pvals

list of p-value cut-offs which will be used to select DEGs

data_names

names of the datasets as they appear in the correlation plot

alphaval

(alpha) transparency of the non-mean boxplots

N_randperms

number of random permutations of the dataset used to select significant DEGs from

N_subsets

number of pairs of random subsets of the dataset used to select significant DEGs from

sex_DEGs

If TRUE, only keep genes present on sex chromosmomes. Queries hspanies gene Ensembl dataset.

fontsize_yaxislabels

font size for axis labels in plot

fontsize_yaxisticks

font size for axis tick labels in plot

fontsize_title

font size for plot title

fontsize_legendlabels

font size for legend labels in plot

fontsize_legendtitle

font size for legend title in plot

fontsize_facet_labels

font size for facet labels

output_path

base path in which outputs will be stored Saves all plots and DGE analysis outputs in the appropriate directories

Examples