Obtain box plots for the correlations of all celltypes, and the mean correlations at a specified cutoff p-value

Obtain box plots for the correlations of all celltypes, and the mean correlations at a specified cutoff p-value

correlation_boxplots(
  corr_mats,
  num_real_datasets,
  pvals = c(0.05, 0.025, 0.01, 0.001, 1e-04),
  alphaval = 0.25,
  N_randperms = 5,
  N_subsets = 5,
  sex_DEGs = FALSE,
  fontsize_yaxislabels = 12,
  fontsize_yaxisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  fontsize_facet_labels = 9,
  output_path = getwd()
)

Arguments

corr_mats

(named) list of correlation matrices for each celltype with the final element being the mean correlation matrix, all at specified p-value

num_real_datasets

total number of real datasets (most likely the number of studies, but sometimes a study may be split e.g. into 2 brain regions, so in this case it would be the number of studies plus 1)

pvals

the cut-off p-value which was used to select DEGs

alphaval

(alpha) transparency of the non-mean boxplots

N_randperms

number of random permutations of the dataset used to select significant DEGs from

N_subsets

number of pairs of random subsets of the dataset used to select significant DEGs from

sex_DEGs

true if DEGs come from sex chromosomes, else false

fontsize_yaxislabels

font size for axis labels in plot

fontsize_yaxisticks

font size for axis tick labels in plot

fontsize_title

font size for plot title

fontsize_legendlabels

font size for legend labels in plot

fontsize_legendtitle

font size for legend title in plot

fontsize_facet_labels

font size for facet labels

output_path

base path in which outputs will be stored Saves box plots for correlation matrices, at a certain p-value cut-off, in the appropriate directory