Tests input parameters for functions

Tests input parameters for functions

validate_input_parameters_bulk(
  SCEs = "placeholder",
  dataset_names = "placeholder",
  celltype = "placeholder",
  celltype_correspondence = "placeholder",
  output_path = "placeholder",
  range_downsampled = "placeholder",
  celltypeIDs = "placeholder",
  sampled = "placeholder",
  sampleIDs = "placeholder",
  bulkDE = "placeholder",
  bulk_cutoff = "placeholder",
  pvalue = "placeholder",
  Nperms = "placeholder",
  fontsize_axislabels = "placeholder",
  fontsize_axisticks = "placeholder",
  fontsize_title = "placeholder",
  fontsize_legendlabels = "placeholder",
  fontsize_legendtitle = "placeholder",
  plot_title = "placeholder"
)

Arguments

SCEs

list of the input data (elements should be SCE objects)

dataset_names

list of the names of the datasets (as you would like them to appear in the "output_path" directory)

celltype

the cell type we are focusing on (name as it appears in cell type sub-directory name)

celltype_correspondence

list of different names specifying each cell type

output_path

path storing the down-sampled DGE analysis for each single-cell dataset, generated for bulk analysis

range_downsampled

vector or list containing values which the data will be downsampled at, in ascending order

celltypeIDs

list or vector of cell type IDs (in order of SCEs)

sampled

downsampling carried out based on what (either "individuals" or "cells")

sampleIDs

list or vector of sample IDs (in order of SCEs)

bulkDE

DGE analysis output for a bulk RNA-seq dataset: rows (rownames) should be the genes, columns should be tissues, and entries should be significance levels

bulk_cutoff

percentage (proportion between 0 and 1), specified so that we select DEGs common across >= bulk_cutoff of the tissues in the Bulk dataset

pvalue

the cut-off p-value used to select DEGs (for both, bulk and scRNA-seq datasets)

Nperms

number of permutations of DGE analysis outputs for each sample

fontsize_axislabels

font size for axis labels in plot

fontsize_axisticks

font size for axis tick labels in plot

fontsize_title

font size for plot title

fontsize_legendlabels

font size for legend labels in plot

fontsize_legendtitle

font size for legend title in plot

plot_title

plot title Checks all bulk analysis parameters are specified correctly