Tests input parameters for functions

Tests input parameters for functions

validate_input_parameters_correlation(
  main_dataset = "placeholder",
  SCEs = "placeholder",
  sampleIDs = "placeholder",
  celltypeIDs = "placeholder",
  celltype_correspondence = "placeholder",
  pvalues = "placeholder",
  dataset_names = "placeholder",
  corr_mats = "placeholder",
  num_real_datasets = "placeholder",
  alphaval = "placeholder",
  N_randperms = "placeholder",
  N_subsets = "placeholder",
  sex_DEGs = "placeholder",
  fontsize_yaxislabels = "placeholder",
  fontsize_yaxisticks = "placeholder",
  fontsize_title = "placeholder",
  fontsize_legendlabels = "placeholder",
  fontsize_legendtitle = "placeholder",
  fontsize_facet_labels = "placeholder",
  output_path = "placeholder"
)

Arguments

main_dataset

name of the dataset used to select significant DEGs from (specified as a string, name as in dataset_names)

SCEs

list of the input data (elements should be SCE objects)

sampleIDs

list or vector of sample IDs (in order of SCEs)

celltypeIDs

list or vector of cell type IDs (in order of SCEs)

celltype_correspondence

list of different names specifying each cell type

pvalues

the list of cut-off p-values which will be used to select DEGs (can just provide a list with one as well)

dataset_names

names of the datasets as they appear in the correlation plot (in order of SCEs)

corr_mats

(named) list of correlation matrices for each celltype with the final element being the mean correlation matrix, all at specified p-value

num_real_datasets

total number of real datasets (most likely the number of studies, but sometimes a study may be split e.g. into 2 brain regions, so in this case it would be the number of studies plus 1)

alphaval

(alpha) transparency of the non-mean boxplots

N_randperms

number of random permutations of the dataset used to select significant DEGs from

N_subsets

number of pairs of random subsets of the dataset used to select significant DEGs from

sex_DEGs

true if DEGs come from sex chromosomes, else false

fontsize_yaxislabels

font size for axis labels in plot

fontsize_yaxisticks

font size for axis tick labels in plot

fontsize_title

font size for plot title

fontsize_legendlabels

font size for legend labels in plot

fontsize_legendtitle

font size for legend title in plot

fontsize_facet_labels

font size for facet labels

output_path

base path in which outputs will be stored Checks all correlation analysis parameters are specified correctly