Validate that there are no user input errors for the differential expression analysis - sc.cell.type.de

Validate that there are no user input errors for the differential expression analysis - sc.cell.type.de

validate_input_parameters_de(
  SCE,
  design,
  sampleID,
  celltypeID,
  y,
  region,
  coef,
  control,
  pval_adjust_method,
  adj_pval,
  output_path,
  rmv_zero_count_genes,
  verbose
)

Arguments

SCE

SingleCellExperiment object, a specialised S4 class for storing data from single-cell experiments

design

Design formula of class type formula. Equation used to fit the model- data for the generalised linear model.

sampleID

Column name in the SCE object to perform pseudobulk on, usually the patient identifier. This column is used for grouping in the pseudobulk approach

celltypeID

Column name in the SCE object for the cell type variable

y

Column name in the SCE object for the return variable e.g. "diagnosis" - Case or disease

region

Column name in the SCE object for the study region. If there are multiple regions in the study (for example two brain regions). Pseudobulk values can be derived separately. Default is "single_region" which will not split by region.

coef

Character specifying which level to investigate for the differential expression analysis e.g. in a case/control study use "case" if case is the identifier in the y column to get positive fold changes to relate to case samples. leave as default value for continuous y. Default is NULL.

control

Character specifying which control level for the differential expression analysis e.g. in a case/control/other study use "control" in the y column to compare against. NOTE only need to specify if more than two groups in y, leave as default value for two groups or continuous y. Default is NULL.

pval_adjust_method

Adjustment method for the p-value in the differential expression analysis. Default is benjamini hochberg "BH". See stats::p.adjust for available options

adj_pval

Adjusted p-value cut-off for the differential expression analysis, 0-1 range

output_path

Folder where the graphs from the differential expression analysis are saved. Default will create a folder in the current working directory "sc.cell.type.de.graphs". False will skip plotting.

rmv_zero_count_genes

Whether genes with no count values in any cell should be removed. Default is TRUE

verbose

Logical indicating if extra information about the differential expression analysis should be printed Checks all DE analysis parameters are specified correctly